checkpoint

+++ project/DepthSensingCUDA/Source/GlobalAppState.h
@@ -103,7 +103,7 @@
 	X(unsigned int, s_streamingThreshold) \
 	X(unsigned int, s_renderFrequency) \
 	X(bool, s_skipFrameEnabled) \
-	X(unsigned int, s_skipFrameThreshold)
+	X(float, s_skipFrameThreshold)
 
 
 #ifndef VAR_NAME
